redis_pg_sync/requirements.txt
# Postgres sync module dependencies
asyncio>=3.4.3
asyncpg>=0.25.0
redis>=4.3.0
python-dotenv>=0.19.0
aioredis>=2.0.0   # For async Redis operations
psycopg2-binary==2.9.7
async-timeout==4.0.3
colorama>=0.4.4